feat(tests): add test hooks loaded before & after orm
There is a difference if a the extensions are loaded before and after the ORM.
The one loaded before will have its models handled by the orm, whereas the one
loaded after the ORM will have to deal with the merger itself. So we need to
test if it works well on both cases.