fix(Hook): fix mistaken hook dir
generally when a module requires another module, Node caches the read file. When the same module is
requested elsewhere it uses the cached module. So working with parent module filename cozes the
first module's dirname to used instead of the actual requiring module dirname. Therefore when the
workaround is to specify the dirname when booting the hook for the specific micro app hook. See the
change.